requirements.txt
flask>=3.0
numpy>=1.24
scipy>=1.10
requests>=2.31
gpxpy>=1.6
trimesh>=4.0
pyproj>=3.6
shapely>=2.0
Pillow>=10.0

# trimesh runtime dependencies that are optional in the base install
# but required for our 3MF export + polygon extrusion pipeline:
networkx>=3.0
mapbox-earcut>=1.0
manifold3d>=2.5
lxml>=5.0
